apps/api/pyproject.toml
[project]
name = "vaultix-api"
version = "0.1.0"
description = "Vaultix FastAPI service"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.8.0",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vaultix_api"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
